airflow/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==1.4.40
xlwt==1.3.0
xlrd==2.0.1
pytest==6.2.5
kafka-python==2.0.2